requirements.txt
pymongo
mysql-connector-python
opencv-python
Pillow
pixelmatch
numpy